Address review on #4981
- Drop the duplicate urlQuery.js tag from /labelMap (ps-map.js already carries it).
- Hoist the role-session helpers shared by AdminLabelPageSpec and ImageryAdminSpec
  into test/util/RoleSession, resolving the minted account through its own
  authenticator instead of diffing sidewalk_user, so concurrent suites can't confuse
  each other's accounts.
- Restate the LabelSearchPage source comment and complete labelPopup's host list.
