+++ app/views/apps/labelMap.scala.html
@@ -34,8 +34,6 @@
   <script src='@assets.path("vendor/i18next/i18next-23.16.8.min.js")'></script>
   <script src='@assets.path("vendor/i18next-http-backend/i18nextHttpBackend-3.0.4.min.js")'></script>
   <script src='@assets.path("vendor/panzoom/panzoom-9.4.4.min.js")'></script>
-  @* Shared deep-link query rules; every reader/writer of the URL's filter params depends on it. *@
-  <script src='@assets.path("js/common/urlQuery.js")'></script>
   @* Label-detail modules (order matters: PopupPanoManager + LabelDetail precede LabelPopup; Toast precedes Badges). *@
   <script src='@assets.path("js/common/aiLabelIndicator.js")'></script>
   <script src='@assets.path("js/common/Toast.js")'></script>
