+++ app/views/common/labelPopup.scala.html
@@ -1,7 +1,8 @@
 @*
 * Modal-popup wrapper around label detail using a native <dialog> with the `.label-detail` styling.
 *
-* Used by: LabelMap, the User Dashboard, the admin Label Map. Each host page also imports the stylesheet (`label-detail.css`) and instantiates the JS controller via `LabelPopup(...)`.
+* Used by: LabelMap, the story list, the User Dashboard and public profile, and the admin Label Map, Activity, and
+* Stories pages. Each host page also imports the stylesheet (`label-detail.css`) and instantiates the JS controller via `LabelPopup(...)`.
 *@
 @(admin: Boolean = false, withPaging: Boolean = false)(implicit assets: AssetsFinder)
